pyproject.toml
[project]
name = "ndviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "nibabel>=5.3.3",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "uvicorn>=0.41.0",
]
